% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license_info.R
\name{license_url}
\alias{license_url}
\title{Create the html tag for linking licenses}
\usage{
license_url(license)
}
\arguments{
\item{license}{A data.frame as generated by \code{licensing_match()}.}
}
\value{
A html tag with the link to the license.
}
\description{
Creates a link to the <choosealicense.com> website.
}
\examples{
license_url(licensing_match("GPL (>= 2.0)"))
}
