% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license_info.R
\name{licensing_match}
\alias{licensing_match}
\title{Find info about a license}
\usage{
licensing_match(license)
}
\arguments{
\item{license}{Character with license info (usually from DESCRIPTION Liencense field)}
}
\value{
A data.frame with name, acronym, several titles and their permissions and the url to the website.
If no is found NULL
}
\description{
Find info about a license
}
\examples{
licensing_match("GPL (>= 2.0)")
}
