% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{package_report}
\alias{package_report}
\title{Reports about a package}
\usage{
package_report(
  package_name,
  package_version,
  package = NULL,
  template_path = system.file("report/pkg_template.qmd", package = "riskreports"),
  params = list(),
  ...
)
}
\arguments{
\item{template_path}{Path to a custom quarto template file}

\item{params}{A list of execute parameters passed to the template}

\item{...}{Additional arguments passed to \code{quarto::quarto_render()}}

\item{x}{Package name or path}
}
\value{
A report
}
\description{
Reports about a package
}
\examples{
package_report(
  package_name = "dplyr",
  package_version = "1.1.4",
  params = list(
    image = "rhub/ref-image",
    assessment_path = "./inst/assessments/dplyr.rds"
  )
)

}
