% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{package_report}
\alias{package_report}
\title{Reports about a package}
\usage{
package_report(x, template_path = NULL, params = list(), ...)
}
\arguments{
\item{x}{Package name or path}

\item{template_path}{Path to a custom quarto template file}

\item{params}{A list of execute parameters passed to the template}

\item{...}{Additional arguments passed to \code{quarto::quarto_render()}}
}
\value{
A report
}
\description{
Reports about a package
}
\examples{
package_report(
  x = "inst/source/dplyr",
  template_path = "inst/report/template.qmd",
  params = list(
    repo = "inst/source",
    package = "dplyr",
    image = "rhub/ref-image",
    assessment_path = "inst/assessments/dplyr.rds"
  )
)

}
