% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{package_report}
\alias{package_report}
\title{Reports about a package}
\usage{
package_report(
  package_name,
  package_version,
  package = NULL,
  template_path = system.file("report/pkg_template.qmd", package = "riskreports"),
  params = list(),
  ...
)
}
\arguments{
\item{package_name}{Package name.}

\item{package_version}{Package version number.}

\item{package}{Path where to find a package source to retrieve name and version number.}

\item{template_path}{Path to a custom quarto template file}

\item{params}{A list of execute parameters passed to the template}

\item{...}{Additional arguments passed to \code{quarto::quarto_render()}}
}
\value{
A path to the reports generated, called by its side effects.
}
\description{
Reports about a package
}
\examples{
pr <- package_report(
  package_name = "dplyr",
  package_version = "1.1.4",
  params = list(
    assessment_path = system.file("assessments/dplyr.rds", package = "riskreports"),
    image = "rhub/ref-image")
)
pr
file.remove(pr)
}
