% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{package_report}
\alias{package_report}
\title{Reports about a package}
\usage{
package_report(x)
}
\arguments{
\item{x}{Package name or path}
}
\value{
A report
}
\description{
Reports about a package
}
\examples{
package_report("inst/source/dplyr_1.1.4/dplyr")


}
